# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZOGO Grocery Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: flashSale
          in: query
          schema:
            type: boolean
        - name: local
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile/addresses:
    get:
      operationId: getAddresses
      tags: [profile]
      summary: Get addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [profile]
      summary: Add address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /profile/addresses/{id}:
    put:
      operationId: updateAddress
      tags: [profile]
      summary: Update address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [profile]
      summary: Delete address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [CUSTOMER, SUPER_ADMIN]
        permissions:
          type: array
          items:
            type: string
            enum:
              - FULL_ACCESS
              - DASHBOARD
              - PRODUCTS
              - ORDERS
              - CUSTOMERS
              - INVENTORY
              - ANALYTICS
              - COUPONS
              - REPORTS
              - SETTINGS
        createdAt:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
      required: [id, name, email, phone, role]
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
    Address:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        name:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required:
        [
          id,
          label,
          name,
          phone,
          line1,
          city,
          district,
          state,
          pincode,
          isDefault,
        ]
    AddressInput:
      type: object
      required: [label, name, phone, line1, city, district, state, pincode]
      properties:
        label:
          type: string
        name:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameBn:
          type: string
        slug:
          type: string
        brand:
          type: string
        category:
          type: string
        categoryId:
          type: string
        description:
          type: string
        weight:
          type: string
        unit:
          type: string
        mrp:
          type: number
        price:
          type: number
        discount:
          type: number
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isFlashSale:
          type: boolean
        isLocal:
          type: boolean
        isOrganic:
          type: boolean
      required:
        [
          id,
          name,
          brand,
          category,
          categoryId,
          mrp,
          price,
          discount,
          stock,
          images,
          rating,
          reviewCount,
        ]
...
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        address:
          $ref: "#/components/schemas/Address"
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        deliveryCharge:
          type: number
        gst:
          type: number
        couponDiscount:
          type: number
        total:
          type: number
        deliverySlot:
          type: string
        estimatedDelivery:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/OrderTimeline"
      required:
        [
          id,
          orderNumber,
          userId,
          items,
          status,
          paymentMethod,
          total,
          createdAt,
        ]
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        image:
          type: string
        quantity:
          type: integer
        price:
          type: number
        totalPrice:
          type: number
      required: [productId, name, image, quantity, price, totalPrice]
    OrderTimeline:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
        message:
          type: string
      required: [status, timestamp, message]
    OrderInput:
      type: object
      required: [addressId, paymentMethod, deliverySlot]
      properties:
        addressId:
          type: string
        paymentMethod:
          type: string
        deliverySlot:
          type: string
        couponCode:
          type: string
        notes:
          type: string
...
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: string
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        discount:
          type: number
        type:
          type: string
        minOrder:
          type: number
        maxDiscount:
          type: number
        description:
          type: string
        isValid:
          type: boolean
      required: [id, code, discount, type, isValid]
    CouponValidateInput:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
        orderTotal:
          type: number
...